Tweak JSON example of unmarshalling a float
The current example shows `json.Unmarshal` on an array of bytes, including
`num:6.0`. In the output however this comes out as `6`. Even in the casting to
`float64`, it comes out as `6`. This can be confusion because it looks like a
simple `int`. By adding an arbitray `.13` you see it's a
float.